chore: make modal package and move modal composable
